Large Update; Windows SDK binding successful
* Solved many issues with naming collisions
* Solved many issues with forward declarations
* Solved many issues with MSVC extensions and keywords
* Added COFF parser to map symbols to their respective libraries
* Updated example.bind with new options
** Added `shallow-bind` instead of `shallow-include` due to issues with type table
** changed `lib` to `libraries` to support linking multiple libraries
* Fixed Bug when tokenizing wide strings and chars (i.e.: L"Wide String")

+++ src/expression.c
@@ -8,7 +8,7 @@
 Token advance_expr(Token_Run *expr)
 {
 	Token prev = *(expr->curr++);
-	while (expr->curr <= expr->end && 
+	while (expr->curr <= expr->end &&
            (expr->curr->kind == Token_Comment
             || expr->curr->kind == Token_BackSlash))
 		expr->curr++;
@@ -23,43 +23,43 @@ int pp_op_precedence(TokenKind k)
         case Token_Quo:
         case Token_Mod:
         return 13;
-        
+
         case Token_Add:
         case Token_Sub:
         return 12;
-        
+
         case Token_Shl:
         case Token_Shr:
         return 11;
-        
+
         case Token_Lt:
         case Token_LtEq:
         case Token_Gt:
         case Token_GtEq:
         return 10;
-        
+
         case Token_CmpEq:
         case Token_NotEq:
         return 9;
-        
+
         case Token_And:
         return 8;
-        
+
         case Token_Xor:
         return 7;
-        
+
         case Token_Or:
         return 6;
-        
+
         case Token_CmpAnd:
         return 5;
-        
+
         case Token_CmpOr:
         return 4;
-        
+
         case Token_Question:
         return 3;
-        
+
         case Token_Eq:
         case Token_AddEq:
         case Token_SubEq:
@@ -72,7 +72,7 @@ int pp_op_precedence(TokenKind k)
         case Token_ShlEq:
         case Token_ShrEq:
         return 2;
-        
+
         default: break;
     }
     return 0;
@@ -81,80 +81,80 @@ int pp_op_precedence(TokenKind k)
 Expr *make_expression(Preprocessor *pp, ExprKind kind)
 {
 	Expr *expr = gb_alloc_item(pp->allocator, Expr);
-    
+
 	expr->kind = kind;
-    
+
 	return expr;
 }
 
 Expr *constant_expr(Preprocessor *pp, u64 val, Constant_Type type, Constant_Format fmt)
 {
 	Expr *expr = make_expression(pp, ExprKind_Constant);
-    
+
 	expr->Constant.val  = val;
 	expr->Constant.type = type;
 	expr->Constant.fmt  = fmt;
-    
+
 	return expr;
 }
 
 Expr *string_expr(Preprocessor *pp, String str)
 {
 	Expr *expr = make_expression(pp, ExprKind_String);
-    
+
 	expr->String.str = str;
-    
+
 	return expr;
 }
 
 Expr *macro_expr(Preprocessor *pp, Token name, Token_Run args)
 {
 	Expr *expr = make_expression(pp, ExprKind_Macro);
-    
+
 	expr->Macro.name = name;
 	expr->Macro.args = args;
-    
+
 	return expr;
 }
 
 Expr *paren_expr(Preprocessor *pp, Expr *sub_expr)
 {
 	Expr *expr = make_expression(pp, ExprKind_Paren);
-    
+
 	expr->Paren.expr = sub_expr;
-    
+
 	return expr;
 }
 
 Expr *unary_expr(Preprocessor *pp, Token op, Expr *operand)
 {
 	Expr *expr = make_expression(pp, ExprKind_Unary);
-    
+
 	expr->Unary.op = op;
 	expr->Unary.operand = operand;
-    
+
 	return expr;
 }
 
 Expr *binary_expr(Preprocessor *pp, Token op, Expr *lhs, Expr *rhs)
 {
 	Expr *expr = make_expression(pp, ExprKind_Binary);
-    
+
 	expr->Binary.op  = op;
 	expr->Binary.lhs = lhs;
 	expr->Binary.rhs = rhs;
-    
+
 	return expr;
 }
 
 Expr *ternary_expr(Preprocessor *pp, Expr *cond, Expr *then, Expr *els_)
 {
 	Expr *expr = make_expression(pp, ExprKind_Ternary);
-    
+
 	expr->Ternary.cond = cond;
 	expr->Ternary.then = then;
 	expr->Ternary.els_ = els_;
-    
+
 	return expr;
 }
 
@@ -163,7 +163,7 @@ String int_string(String str, Constant_Type *type, Constant_Format *format)
 {
     String num;
     char *curr = str.start;
-    
+
     num.start = curr;
     format->base = 10;
     if (gb_strncmp(curr, "0x", 2) == 0)
@@ -181,7 +181,7 @@ String int_string(String str, Constant_Type *type, Constant_Format *format)
         curr++;
         format->base = 8;
     }
-    
+
     format->sig_figs = 0;
     while ((format->base == 2  && gb_is_between(curr[0], '0', '1')) ||
            (format->base == 8  && gb_is_between(curr[0], '0', '7')) ||
@@ -191,11 +191,11 @@ String int_string(String str, Constant_Type *type, Constant_Format *format)
         format->sig_figs++;
         curr++;
     }
-    
+
     num.len = curr - num.start;
     if (cstring_cmp(num, "0") == 0 && format->base == 8)
         format->base = 10;
-    
+
     type->size = Constant_32;
     type->is_signed = true;
     for (;;)
@@ -214,7 +214,7 @@ String int_string(String str, Constant_Type *type, Constant_Format *format)
 u64 char_lit_val(Token lit)
 {
     GB_ASSERT(lit.kind == Token_Char || lit.kind == Token_Wchar);
-    
+
     String lit_str = lit.str;
     switch (lit.kind)
     {
@@ -227,7 +227,7 @@ u64 char_lit_val(Token lit)
         default:
         break;
     }
-    
+
     const u8 charconsts[] = {7, 8, 27, 12, 10, 13, 9, 11};
     if (lit_str.start[0] == '\\')
     {
@@ -236,9 +236,9 @@ u64 char_lit_val(Token lit)
             case '0': case '1': case '2': case '3':
             case '4': case '5': case '6': case '7':
             return str_to_int_base(string_slice(lit_str, 1, lit_str.len), 8);
-            
+
             case 'x': return str_to_int_base(string_slice(lit_str, 2, lit_str.len), 16);
-            
+
             case 'a': return charconsts[0];
             case 'b': return charconsts[1];
             case 'e':
@@ -262,7 +262,7 @@ Expr *pp_parse_operand(Token_Run *expr, Preprocessor *pp)
 {
 	Constant_Type type = DEFAULT_TYPE;
 	Constant_Format fmt = DEFAULT_FORMAT;
-    
+
 	switch (expr->curr->kind)
 	{
         case Token_Integer: {
@@ -270,7 +270,7 @@ Expr *pp_parse_operand(Token_Run *expr, Preprocessor *pp)
             advance_expr(expr);
             return constant_expr(pp, val, type, fmt);
         }
-        
+
         case Token_Char: {
             type.is_signed = false;
             type.size = sizeof(char);
@@ -279,7 +279,7 @@ Expr *pp_parse_operand(Token_Run *expr, Preprocessor *pp)
             advance_expr(expr);
             return constant_expr(pp, val, type, fmt);
         }
-        
+
         case Token_Wchar: {
             type.is_signed = false;
             type.size = sizeof(wchar_t);
@@ -288,14 +288,14 @@ Expr *pp_parse_operand(Token_Run *expr, Preprocessor *pp)
             advance_expr(expr);
             return constant_expr(pp, val, type, fmt);
         }
-        
+
         case Token_String: {
             String str = expr->curr->str;
             advance_expr(expr);
             return string_expr(pp, str);
         }
-        
-        default: 
+
+        default:
 		if (!gb_is_between(expr->curr->kind, Token__KeywordBegin, Token__KeywordEnd))
 			break;
         case Token_Ident: {
@@ -316,7 +316,7 @@ Expr *pp_parse_operand(Token_Run *expr, Preprocessor *pp)
             }
             return macro_expr(pp, name, args);
         }
-        
+
         case Token_OpenParen: {
             advance_expr(expr);
             Expr *expression = paren_expr(pp, _pp_parse_expression(expr, pp));
@@ -346,10 +346,10 @@ Expr *pp_parse_unary_expression(Token_Run *expr, Preprocessor *pp)
             Token op = advance_expr(expr);
             return unary_expr(pp, op, pp_parse_unary_expression(expr, pp));
         }
-        
+
         default: break;
 	}
-    
+
 	return pp_parse_operand(expr, pp);
 }
 Expr *pp_parse_ternary_expr(Token_Run *expr, Preprocessor *pp, Expr *cond)
@@ -358,10 +358,10 @@ Expr *pp_parse_ternary_expr(Token_Run *expr, Preprocessor *pp, Expr *cond)
 	then = _pp_parse_expression(expr, pp);
 	Token colon = advance_expr(expr);
 	els_ = _pp_parse_expression(expr, pp);
-    
+
 	if (colon.kind != Token_Colon)
 		syntax_error(colon, "Expected ':', got '%.*s'", LIT(TokenKind_Strings[colon.kind]));
-    
+
 	return ternary_expr(pp, cond, then, els_);
 }
 Expr *pp_parse_binary_expr(Token_Run *expr, Preprocessor *pp, int max_prec)
@@ -378,7 +378,7 @@ Expr *pp_parse_binary_expr(Token_Run *expr, Preprocessor *pp, int max_prec)
 			if (op_prec != prec) break;
 			if (op_prec == 0) error(op, "Expected operator, got '%.*s'", LIT(TokenKind_Strings[expr->curr->kind]));
 			advance_expr(expr);
-            
+
 			if (op.kind == Token_Question)
 			{
 				expression = pp_parse_ternary_expr(expr, pp, expression);
@@ -392,7 +392,7 @@ Expr *pp_parse_binary_expr(Token_Run *expr, Preprocessor *pp, int max_prec)
 			}
 		}
 	}
-    
+
 	return expression;
 }
 
@@ -422,7 +422,7 @@ b32 num_greater_eq(Expr_Constant lhs, Expr_Constant rhs)
         if (is_pos != num_positive(rhs))
             return is_pos;
     }
-    
+
     return (lhs.val >= rhs.val);
 }
 
@@ -430,7 +430,7 @@ Constant_Type convert_int_types(Constant_Type *lhs, Constant_Type *rhs)
 {
     if (lhs->size < sizeof(int)) lhs->size = sizeof(int);
     if (rhs->size < sizeof(int)) rhs->size = sizeof(int);
-    
+
     if (lhs->size == rhs->size && lhs->is_signed == rhs->is_signed) return *lhs;
     if (lhs->size == rhs->size)
     {
@@ -454,10 +454,10 @@ u64 _pp_expr_defined(Preprocessor *pp, Expr *expr)
 {
 	while (expr->kind == ExprKind_Paren)
 		expr = expr->Paren.expr;
-    
+
 	if (expr->kind != ExprKind_Macro)
 		gb_printf_err("ERROR: Operand of 'defined' is not an identifer\n");
-    
+
 	Define def = pp_get_define(pp, expr->Macro.name.str);
 	return def.in_use;
 }
@@ -470,7 +470,7 @@ Expr *_pp_expand_macro_expr(Preprocessor *pp, Expr *expr)
 		return constant_expr(pp, 0, DEFAULT_TYPE, DEFAULT_FORMAT);
     //error(expr->Macro.name, "Macro '%.*s' is not defined", LIT(expr->Macro.name.str));
     //gb_printf_err("ERROR: Macro '%.*s' is not defined\n", LIT(expr->Macro.name.str));
-    
+
 	Expr *parsed = 0;
 	if (def.params)
 	{
@@ -485,7 +485,7 @@ Expr *_pp_expand_macro_expr(Preprocessor *pp, Expr *expr)
 		Token_Run value = def.value;
 		parsed = _pp_parse_expression(&value, pp);
 	}
-    
+
     return _pp_eval_expression(pp, parsed);
 }
 
@@ -496,11 +496,11 @@ Expr *_pp_eval_expression(Preprocessor *pp, Expr *expr)
         case ExprKind_Constant:
         return gb_alloc_copy(gb_heap_allocator(), expr, sizeof(Expr));
         break;
-        
+
         case ExprKind_Macro: {
             return _pp_expand_macro_expr(pp, expr);
         }
-        
+
         case ExprKind_Unary: {
             if (expr->Unary.op.kind == Token_Ident &&
                 cstring_cmp(expr->Unary.op.str, "defined") == 0)
@@ -509,20 +509,20 @@ Expr *_pp_eval_expression(Preprocessor *pp, Expr *expr)
                 u64 res = _pp_expr_defined(pp, expr->Unary.operand);
                 return constant_expr(pp, res, DEFAULT_TYPE, DEFAULT_FORMAT);
             }
-            
+
             Expr *operand_expr = _pp_eval_expression(pp, expr->Unary.operand);
-            
+
             Expr_Constant operand = operand_expr->Constant;
             u64 res;
             switch (expr->Unary.op.kind)
             {
                 case Token_Not: res = !operand.val; break;
-                case Token_Sub: res = -operand.val; break;
+                case Token_Sub: res = -(i64)operand.val; break;
                 default: res = 0; break;
             }
             return constant_expr(pp, res, DEFAULT_TYPE, DEFAULT_FORMAT);
         } break;
-        
+
         case ExprKind_Paren: {
             return _pp_eval_expression(pp, expr->Paren.expr);
         }
@@ -531,10 +531,10 @@ Expr *_pp_eval_expression(Preprocessor *pp, Expr *expr)
             Expr *rhs_expr = _pp_eval_expression(pp, expr->Binary.rhs);
             Expr_Constant lhs = lhs_expr->Constant;
             Expr_Constant rhs = rhs_expr->Constant;
-            
+
             u64 res;
             Constant_Type result_type = convert_int_types(&lhs.type, &rhs.type);
-            
+
             switch (expr->Binary.op.kind)
             {
                 case Token_Add   : res = lhs.val +  rhs.val; break;
@@ -558,16 +558,16 @@ Expr *_pp_eval_expression(Preprocessor *pp, Expr *expr)
             }
             return constant_expr(pp, res, result_type, DEFAULT_FORMAT);
         }
-        
+
         case ExprKind_Ternary: {
             Expr *cond = _pp_eval_expression(pp, expr->Ternary.cond);
             Expr *then = _pp_eval_expression(pp, expr->Ternary.then);
             Expr *els_ = _pp_eval_expression(pp, expr->Ternary.els_);
-            
+
             u64 cond_val = cond->Constant.val;
             return cond_val ? then : els_;
         }
-        
+
         default:
     	gb_printf_err("ERROR: Invalid Expression type in pp_eval_expression\n");
         return constant_expr(pp, 0, DEFAULT_TYPE, DEFAULT_FORMAT);
@@ -582,5 +582,5 @@ u64 pp_eval_expression(Preprocessor *pp, Expr *expr)
 
 void free_expr(gbAllocator a, Expr *expr)
 {
-    
+
 }
